tools/pipeline-builder/pipeline_builder_lambda/requirements.txt
boto3
gitpython
cookiecutter